refactor: emoji-data files for better visibility
