+++ docusaurus/docs/reactnative/guides/auto_complete_suggestion_list_customizations.mdx
@@ -164,7 +164,7 @@ Note: The default `AutoCompleteSuggestionList` must be returned if you don't wan
 
 ### Customize Auto complete emoji search result
 
-Our `AutoCompleteSuggestionList` already support emoji search using a query that starts with `:` in the message input, on a default data set that can be found [here](https://github.com/GetStream/stream-chat-react-native/blob/develop/package/src/emoji-data/compiled.ts).
+Our `AutoCompleteSuggestionList` already support emoji search using a query that starts with `:` in the message input, on a default data set that can be found [here](https://github.com/GetStream/stream-chat-react-native/blob/develop/package/src/emoji-data/index.ts).
 
 We allow overriding this behaviour and the default search behaviour by using the `emojiSearchIndex` prop in the `Channel` component.
