+++ package/src/emoji-data/index.ts
