+++ package/src/index.ts
@@ -2,7 +2,7 @@ export * from './components';
 export * from './hooks';
 export { registerNativeHandlers, NetInfo } from './native';
 export * from './contexts';
-export * from './emoji-data/compiled';
+export * from './emoji-data';
 
 export * from './icons';
