+++ package/src/utils/utils.ts
@@ -23,7 +23,7 @@ import type {
   SuggestionComponentType,
   SuggestionUser,
 } from '../contexts/suggestionsContext/SuggestionsContext';
-import { compiledEmojis, Emoji } from '../emoji-data/compiled';
+import { compiledEmojis, Emoji } from '../emoji-data';
 import type { IconProps } from '../icons/utils/base';
 import type { TableRowJoinedUser } from '../store/types';
 import type { DefaultStreamChatGenerics, ValueOf } from '../types/types';
